crud and auth done

+++ src/tasks/tasks.service.ts
@@ -2,11 +2,11 @@ import { Injectable,NotFoundException } from '@nestjs/common';
 // import { Task } from './task.model';
 import {TaskStatus} from './task-status.enum'
 import { CreateTaskDto } from './dto/create-task.dto'
-import { v4 as uuidv4 }from 'uuid';
 import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { GetTaskFilterDto } from './dto/get-task-filteer.dto';
+import { User } from 'src/auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -14,11 +14,17 @@ export class TasksService {
     @InjectRepository(TaskRepository)
     private taskRepository: TaskRepository,
   ){}
-  async getTask(fiterDto: GetTaskFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTask(fiterDto)
+  async getTask(
+    fiterDto: GetTaskFilterDto,
+    user: User
+    ): Promise<Task[]> {
+    return this.taskRepository.getTask(fiterDto, user)
   }
-  async getTaskById(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskById(
+    id: number,
+    user: User,
+    ): Promise<Task> {
+    const found = await this.taskRepository.findOne({ where: {id, userId: user.id} });
 
     if(!found){
       throw new NotFoundException('Task does not exist')
@@ -26,17 +32,27 @@ export class TasksService {
 
     return found;
   }
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto)
+  async createTask(
+    createTaskDto: CreateTaskDto,
+    user: User,
+  ): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user)
   }
-  async deleteTask(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id)
+  async deleteTask(
+    id: number,
+    user: User,
+    ): Promise<void> {
+      const result = await this.taskRepository.delete({id, userId: user.id})
     if(result.affected === 0){
       throw new NotFoundException('Task does not exist')
     }
   }
-  async updateTaskStatus(id: number, status: TaskStatus) {
-    const task = await this.getTaskById(id)
+  async updateTaskStatus(
+    id: number,
+    status: TaskStatus,
+    user: User
+    ): Promise<Task> {
+    const task = await this.getTaskById(id, user)
     task.status = status
     await task.save()
